renamed authenticator to passport and modified main to use require that file.
